requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
bottleneck=1.3.4=py39h080aedc_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2022.07.19=haa95532_0
certifi=2022.6.15=py39haa95532_0
cffi=1.15.0=py39h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.4=pyh9f0ad1d_0
configparser=5.0.2=pyhd3eb1b0_0
cryptography=36.0.0=py39h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
et_xmlfile=1.1.0=py39haa95532_0
fftw=3.3.9=h2bbff1b_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
h5py=3.6.0=py39h3de5c98_0
hdf5=1.10.6=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_0
kiwisolver=1.3.2=py39hd77b12b_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
matplotlib=3.5.1=py39haa95532_1
matplotlib-base=3.5.1=py39hd77b12b_1
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
munkres=1.1.4=py_0
numexpr=2.8.1=py39hb80d3ca_0
numpy=1.21.5=py39h7a0a035_1
numpy-base=1.21.5=py39hca35cd5_1
opencv-python=4.5.5.64=pypi_0
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.1.1q=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39hd77b12b_0
pillow=9.0.1=py39hdc2b20a_0
pip=21.2.4=py39haa95532_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39hd77b12b_6
pyreadline=2.1=py39haa95532_1
pysocks=1.7.1=py39haa95532_0
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda11.3_cudnn8_0
pytorch-model-summary=0.1.1=py_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
qt=5.9.7=vc14h73c81de_0
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.1.1=py39hd77b12b_0
scipy=1.7.3=py39h7a0a035_2
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=61.2.0=py39haa95532_0
sip=4.19.13=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=h2bbff1b_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h2bbff1b_0
torch-summary=1.4.5=pypi_0
torchaudio=0.11.0=py39_cu113
torchvision=0.12.0=py39_cu113
tornado=6.1=py39h2bbff1b_0
tqdm=4.64.0=pyhd8ed1ab_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
xz=5.2.5=h62dcd97_0
zlib=1.2.12=h8cc25b3_2
zstd=1.4.9=h19a0ad4_0
